requirement.txt
Django==2.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.10.1
jieba==0.39
Markdown==3.1.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.1
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
Whoosh==2.7.4
